Seems simpler to just have the partitioning in crossvalidation itself
